requirements.txt
dvc==3.55.2
dvc-gdrive==3.0.1
fastapi==0.115.0
imbalanced-learn==0.12.3
joblib==1.4.2
mlflow==2.16.2
mkdocs==1.6.1
mkdocs-material==9.5.38
numpy==2.1.1
pandas==2.2.3
pytest==8.3.3
PyYAML==6.0.2
scikit-learn==1.5.2
setuptools==75.1.0
uvicorn==0.31.0
# evidently @ git+https://github.com/evidentlyai/evidently.git